refactor trace writer externalized blob helpers
